添加分类查询；

+++ src/Dotnet9.Web.Go/models/home.go
@@ -6,7 +6,7 @@ type HomeResponse struct {
 	config.Viewer
 	Categories []Category
 	Posts []BlogPostMore
-	Total int
+	Total int64
 	Page int
 	Pages []int
 	PageEnd bool
